chore: Move breakPointCoords back. Clean up useEffect
